+++ sdk/dns/arm-dns/src/classic/dnssecConfigs/index.ts
@@ -11,6 +11,7 @@ import {
 } from "../../api/dnssecConfigs/options.js";
 import { DnssecConfig } from "../../models/models.js";
 import { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
+import { SimplePollerLike, getSimplePoller } from "../../static-helpers/simplePollerHelpers.js";
 import { PollerLike, OperationState } from "@azure/core-lro";
 
 /** Interface representing a DnssecConfigs operations. */
@@ -27,12 +28,36 @@ export interface DnssecConfigsOperations {
     zoneName: string,
     options?: DnssecConfigsDeleteOptionalParams,
   ) => PollerLike<OperationState<void>, void>;
+  /** @deprecated use delete instead */
+  beginDelete: (
+    resourceGroupName: string,
+    zoneName: string,
+    options?: DnssecConfigsDeleteOptionalParams,
+  ) => Promise<SimplePollerLike<OperationState<void>, void>>;
+  /** @deprecated use delete instead */
+  beginDeleteAndWait: (
+    resourceGroupName: string,
+    zoneName: string,
+    options?: DnssecConfigsDeleteOptionalParams,
+  ) => Promise<void>;
   /** Creates or updates the DNSSEC configuration on a DNS zone. */
   createOrUpdate: (
     resourceGroupName: string,
     zoneName: string,
     options?: DnssecConfigsCreateOrUpdateOptionalParams,
   ) => PollerLike<OperationState<DnssecConfig>, DnssecConfig>;
+  /** @deprecated use createOrUpdate instead */
+  beginCreateOrUpdate: (
+    resourceGroupName: string,
+    zoneName: string,
+    options?: DnssecConfigsCreateOrUpdateOptionalParams,
+  ) => Promise<SimplePollerLike<OperationState<DnssecConfig>, DnssecConfig>>;
+  /** @deprecated use createOrUpdate instead */
+  beginCreateOrUpdateAndWait: (
+    resourceGroupName: string,
+    zoneName: string,
+    options?: DnssecConfigsCreateOrUpdateOptionalParams,
+  ) => Promise<DnssecConfig>;
   /** Gets the DNSSEC configuration. */
   get: (
     resourceGroupName: string,
@@ -53,11 +78,43 @@ function _getDnssecConfigs(context: DnsManagementContext) {
       zoneName: string,
       options?: DnssecConfigsDeleteOptionalParams,
     ) => $delete(context, resourceGroupName, zoneName, options),
+    beginDelete: async (
+      resourceGroupName: string,
+      zoneName: string,
+      options?: DnssecConfigsDeleteOptionalParams,
+    ) => {
+      const poller = $delete(context, resourceGroupName, zoneName, options);
+      await poller.submitted();
+      return getSimplePoller(poller);
+    },
+    beginDeleteAndWait: async (
+      resourceGroupName: string,
+      zoneName: string,
+      options?: DnssecConfigsDeleteOptionalParams,
+    ) => {
+      return await $delete(context, resourceGroupName, zoneName, options);
+    },
     createOrUpdate: (
       resourceGroupName: string,
       zoneName: string,
       options?: DnssecConfigsCreateOrUpdateOptionalParams,
     ) => createOrUpdate(context, resourceGroupName, zoneName, options),
+    beginCreateOrUpdate: async (
+      resourceGroupName: string,
+      zoneName: string,
+      options?: DnssecConfigsCreateOrUpdateOptionalParams,
+    ) => {
+      const poller = createOrUpdate(context, resourceGroupName, zoneName, options);
+      await poller.submitted();
+      return getSimplePoller(poller);
+    },
+    beginCreateOrUpdateAndWait: async (
+      resourceGroupName: string,
+      zoneName: string,
+      options?: DnssecConfigsCreateOrUpdateOptionalParams,
+    ) => {
+      return await createOrUpdate(context, resourceGroupName, zoneName, options);
+    },
     get: (resourceGroupName: string, zoneName: string, options?: DnssecConfigsGetOptionalParams) =>
       get(context, resourceGroupName, zoneName, options),
   };